Fixes association and other bits
Reviewing the association I noticed a few things were not being accounted for correctly:

* New sources were not being added in and given a cat id in each iteration.
* Duplicate matches not being accounted for.
* The first catalogue was not being fully integrated and used as the base 'sources'.
* New sources were not being flagged correctly.

The above problems meant that new sources were not being injected so they were missing from the final sources (I noticed this by a new source that definitely appeared not being there in the test run). It also caused what I called 'stranded sources'. There were some new sources that did make it in because they had been crossmatched with a previous source, but that previous source was already assigned to a different crossmatch. So while the stranded source appeared in the sources list, it's crossmatch was not associated to it, so it appeared alone.

This commit fixes the above issues, and through testing seems to perform the association correctly.

In this basic form of association, double matches are accounted for but not flagged in any way. I keep only the closest match of a duplicate and move the other match(es) to be a 'new' source by itself. I have an idea of a why to improve this (TraP methodology would also account for this.)

New source flagging is also somewhat fixed but needs sky region information to be fully correct. It will not be correct when different sky regions are processed at once.

Finally this also switches to using the weighted average to calculate the average RA and Dec for association during the iterations.