fix "back to top" TOC link
The code to add a "back to top" link to the table of contents would fail with an error message on pages that don't have a TOC.
The link was also never added if the user started on a page without a TOC as the mkdocs-material instant loading feature prevents JS from being executed when navigating between local docs pages. This patch ensures the link is always added.