Add debug logs showing min,max,median values derived from the RMS image
NOTE: I'm not using f-strings for log messages because this is considered bad
practice in logging. f-strings are greedy and will be computed even if the log
isn't printed to the screen. While this isn't an issue here since the log
message is quick to render, it's good practice to always print log messages
in this way in case the message requires an expensive calculation.
See: https://bugs.python.org/issue46200 for some discussion of this.